REF : components 분리 및 textField속성 singleLine ->  maxLine 변경
